requirements.txt
tensorflow>=2.10.0,<2.11.0
numpy>=1.23.0,<1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
jupyter>=1.0.0
Pillow>=9.5.0
opencv-python>=4.7.0
gradio>=4.0.0
